requirements.txt
flask
python-dotenv
requests
beautifulsoup4
selenium
undetected-chromedriver
pytest
werkzeug
typing-extensions
mimetypes-detect
uuid